Indeed-jobs/requirements.txt
# Indeed (JS-rendered pages)
playwright>=1.40.0

# Config
python-dotenv>=1.0.0

# AI: Gemini (use one; code supports both)
google-genai>=1.0.0
google-generativeai>=0.8.0
# OpenAI (optional; set JAR_AI_PROVIDER=openai and OPENAI_API_KEY)
openai>=1.0.0

# Markdown → HTML for PDF export
markdown>=3.5.0
